+++ components/organisms/Dashboard/Layouts/index.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+import Content from "../Content";
+import Sidebar from "../Sidebar";
+
+type MainLayoutsProps = { children: ReactNode }
+
+export default function MainLayouts({ children }: MainLayoutsProps) {
+    return (
+        <>
+            <Sidebar />
+            <Content content={children} />
+        </>
+    )
+}