+++ src/Dashboard/index.tsx
@@ -1,0 +1,13 @@
+import Data from "../../components/organisms/Dashboard/Data";
+import Header from "../../components/organisms/Dashboard/Header";
+import Insights from "../../components/organisms/Dashboard/Insights";
+
+export default function Dashboard() {
+    return (
+        <>
+            <Header />
+            <Insights />
+            <Data />
+        </>
+    )
+}