+++ src/Product/index.tsx
@@ -1,0 +1,5 @@
+export default function Product() {
+    return (
+        <div style={{ color: 'var(--dark)' }}>Product sanryuu</div>
+    )
+}